+++ apps/strapi-cms/src/utils/validateProductPresence.ts
@@ -1,4 +1,5 @@
 import { errors } from '@strapi/utils';
+import { GUIDES_UID, RELEASE_NOTES_UID } from '../middlewares/documentHooks';
 
 const validatorsAreDisabled = process.env.DISABLE_CUSTOM_VALIDATORS === 'True';
 
@@ -24,6 +25,20 @@ export interface IEventWithProduct {
     readonly locale?: string;
   };
 }
+
+export const entitiesRequiringProductAssociation = [
+  'api::api-data-list-page.api-data-list-page',
+  'api::api-data.api-data',
+  GUIDES_UID,
+  'api::guide-list-page.guide-list-page',
+  'api::overview.overview',
+  'api::quickstart-guide.quickstart-guide',
+  'api::tutorial-list-page.tutorial-list-page',
+  'api::tutorial.tutorial',
+  'api::use-case-list-page.use-case-list-page',
+  'api::use-case.use-case',
+  RELEASE_NOTES_UID,
+];
 
 export const validateAssociatedProductPresenceOnUpdate = (
   event: IEventWithProduct
